+++ data/tasks.txt
@@ -4,3 +4,6 @@ D|0| |21 Aug 2021 1200
 D|0|return boo|21 Aug 2021 1012
 D|0|return book|21 Aug 2021 1200
 D|0|return book |21 Aug 2021 1200
+T|0|return book
+D|0|return book|21 Jan 2021 1200
+E|0|meeting|21 Feb 2021 1000
