+++ src/main/java/duke/Duke.java
@@ -1,0 +1,45 @@
+package duke;
+
+import duke.commands.Command;
+
+import java.io.IOException;
+
+public class Duke {
+    private static final String FILE_DIR = "data";
+    private static final String FILE_NAME = "duke.txt";
+
+    private TaskList tasks;
+    private Ui ui;
+    private Storage storage;
+
+    public Duke(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+
+        try {
+            tasks = new TaskList(storage.load());
+        } catch (IOException e) {
+            ui.printError(e.getMessage());
+        }
+    }
+
+    public void run() {
+        ui.printWelcomeMessage();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                Command c = Parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+
+            } catch (DukeException e) {
+                ui.printError(e.getMessage());
+            }
+        }
+    }
+
+    public static void main(String[] args) {
+        new Duke("data/tasks.txt").run();
+    }
+}
