+++ src/main/java/duke/TaskList.java
@@ -1,6 +1,8 @@
-import java.time.LocalDateTime;
+package duke;
+
+import duke.tasks.Task;
+
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 
 public class TaskList {
