+++ src/main/java/duke/commands/DeadlineCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.tasks.Deadline;
+
 import java.io.IOException;
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
