+++ src/main/java/duke/commands/DoneCommand.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 import java.io.IOException;
 
 public class DoneCommand extends Command {
