+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,4 +1,10 @@
-public class ExitCommand extends Command{
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ExitCommand extends Command {
     public static final String COMMAND_WORD = "bye";
     public static final String MESSAGE_USAGE = COMMAND_WORD + " - exit this program";
